Refactor server and db files for organization.
