+++ server/config/db-config.js
@@ -7,31 +7,16 @@ var config = {
 
 var db = new Sequelize('Scavenger', config.username, config.password);
 
-var User = db.define('user', {
-  username: Sequelize.STRING,
-  password: Sequelize.STRING
-});
-
-var Game = db.define('game', {
-  creatorId: Sequelize.INTEGER,
-  path: Sequelize.STRING
-});
-
-var Location = db.define('location', {
-  sequence: Sequelize.INTEGER,
-  lat: Sequelize.DOUBLE,
-  lng: Sequelize.DOUBLE
-});
-
-var LocStatus = db.define('locStatus', {
-  status: Sequelize.BOOLEAN
-});
+var User = require('../games/games')(db);
+var Locations = require('../locations/locations')(db);
+var Status = require('../status/status')(db);
+var User = require('../users/users')(db);
 
 User.belongsToMany(Game, {through: UserGame});
 Game.belongsToMany(User, {through: UserGame});
 
-Location.belongsToMany(User, {through: LocStatus});
-User.belongsToMany(Location, {through: LocStatus});
+Location.belongsToMany(User, {through: Status});
+User.belongsToMany(Location, {through: Status});
 
 Game.belongsTo(User, {as: creatorId});
