+++ server/games/games.js
@@ -1,0 +1,8 @@
+module.exports = function(db) {
+  var Game = db.define('game', {
+    creatorId: Sequelize.INTEGER,
+    path: Sequelize.STRING
+  });
+
+  return Game;
+};
