+++ server/status/status.js
@@ -1,0 +1,7 @@
+module.exports = function(db) {
+  var Status = db.define('Status', {
+    status: Sequelize.BOOLEAN
+  });
+
+  return Status;
+};
