+++ server/users/users.js
@@ -1,0 +1,8 @@
+module.exports = function(db) {
+  var User = db.define('user', {
+    username: Sequelize.STRING,
+    password: Sequelize.STRING
+  });
+
+  return User;
+};
